requirements.txt
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
Pillow==9.4.0